+++ packages/gits_extension/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4
+
+- update package time: ^2.1.3
+
 ## 0.0.3
 
 - update package time to 2.1.2